FIXTURE: Improve project structure
