+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines": {"/home/flavia/Desktop/week1_andela/shopping_list/tests/test_shoppinglist.py": [1, 2, 3, 4, 5, 6, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 28, 29, 30, 31, 33, 35, 36, 37, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 53, 55, 56, 57, 58, 60, 62, 63, 64, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 78, 80, 81, 82, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 100, 102, 103, 104, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 126, 127, 128, 129, 131, 133, 134, 135, 137, 139, 140, 141, 142, 143, 144, 145, 146, 148, 150, 151, 152, 154, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 167, 169, 170, 171, 173, 174, 175, 176, 177, 178, 179, 180, 181], "/home/flavia/Desktop/week1_andela/shopping_list/tests/test_items.py": [1, 2, 3, 4, 5, 6, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 22, 23, 24, 26, 27, 28, 29, 31, 33, 34, 35, 37, 38, 39, 40, 41, 43, 44, 45, 46, 47, 48, 50, 52, 53, 54, 56, 58, 59, 60, 61, 62, 63, 64, 66, 67, 68, 70, 71, 72, 73, 74, 75, 77, 78, 79, 82, 83, 84, 85, 87, 88, 90, 92, 93, 94, 96, 97, 98, 99, 100, 101, 103, 104, 105, 106, 107, 108, 109, 111, 112, 113, 115, 116, 117, 118, 120, 121, 122, 123, 125, 126, 127, 128, 129, 130, 132, 133, 134, 135, 136, 137, 138, 139, 142, 143, 144, 145, 146, 147, 149, 150, 151], "/home/flavia/Desktop/week1_andela/shopping_list/tests/test_user.py": [1, 2, 3, 4, 5, 6, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 20, 21, 22, 24, 26, 27, 28, 30, 32, 33, 34, 35, 36, 38, 40, 41, 42, 43, 44, 45, 46, 47, 48, 50, 52, 53, 54, 55, 56, 57, 58, 60, 62, 63, 64, 65, 66, 67, 68, 70, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 85, 87, 88, 89, 90, 91, 92, 93, 94]}}
+!coverage.py: This is a private format, don't read it directly!{"lines": {"/Users/flavianshem/Projects/shopping_list/app/__init__.py": [1, 2, 3, 5, 7, 8, 9, 10, 11, 15, 16, 17, 20, 21, 22, 24], "/Users/flavianshem/Projects/shopping_list/app/views.py": [1, 2, 3, 4, 5, 7, 22, 46, 63, 116, 142, 171, 194, 220, 260, 295, 10, 12, 13, 14, 15, 16, 18, 25, 27, 30, 31, 34, 36, 37, 39, 40, 42, 197, 198, 200, 201, 203, 204, 205, 67, 68, 70, 71, 73, 74, 75, 76, 77, 78, 79, 206, 207, 208, 209, 210, 211, 299, 300, 302, 303, 304, 305, 307, 308, 309, 310, 311, 312, 223, 224, 226, 227, 229, 230, 231, 232, 233, 234, 236, 237, 238, 239, 241, 246, 248, 250, 251, 264, 265, 267, 268, 269, 271, 272, 273, 275, 276, 284, 285, 277, 278, 279, 280, 281, 282, 283, 286, 81, 82, 114, 175, 176, 178, 179, 181, 182, 183, 184, 185, 146, 147, 149, 150, 152, 153, 154, 155, 157, 158, 160, 162, 163, 85, 86, 87, 89, 92, 94, 95, 102, 103, 106, 107, 96, 97, 98, 99, 100, 108, 90, 104, 120, 121, 123, 124, 126, 128, 134, 129, 130, 131, 132, 133, 49, 51, 52, 53, 56, 57, 58, 60, 54, 19, 43, 32], "/Users/flavianshem/Projects/shopping_list/app/models.py": [1, 2, 3, 4, 6, 7, 8, 9, 10, 11, 13, 18, 23, 28, 45, 59, 63, 64, 65, 66, 67, 68, 69, 70, 72, 77, 80, 85, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 102, 109, 112, 117, 14, 15, 16, 20, 21, 30, 33, 34, 35, 37, 38, 39, 40, 48, 50, 51, 73, 74, 75, 82, 83, 103, 104, 105, 106, 107, 114, 115, 119, 120, 87, 88], "/Users/flavianshem/Projects/shopping_list/app/tests/__init__.py": [1], "/Users/flavianshem/Projects/shopping_list/app/tests/test_items.py": [1, 2, 3, 4, 5, 6, 7, 10, 11, 12, 23, 27, 32, 51, 91, 121, 13, 14, 15, 16, 17, 18, 19, 20, 21, 34, 35, 36, 28, 29, 30, 38, 39, 40, 41, 42, 44, 45, 46, 47, 48, 49, 24, 25, 122, 123, 124, 126, 127, 128, 129, 130, 131, 133, 134, 135, 136, 137, 138, 139, 140, 143, 144, 145, 146, 147, 148, 150, 151, 152, 53, 54, 55, 57, 59, 60, 61, 62, 63, 64, 65, 67, 68, 69, 71, 72, 73, 74, 75, 76, 78, 79, 80, 83, 84, 85, 86, 88, 89, 93, 94, 95, 97, 98, 99, 100, 101, 102, 104, 105, 106, 107, 108, 109, 110, 112, 113, 114, 116, 117, 118, 119], "/Users/flavianshem/Projects/shopping_list/app/tests/test_shoppinglist.py": [1, 2, 3, 4, 5, 6, 9, 10, 11, 24, 28, 33, 53, 60, 78, 100, 126, 131, 148, 167, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 35, 36, 37, 29, 30, 31, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 25, 26, 55, 56, 57, 58, 169, 170, 171, 173, 174, 175, 176, 177, 178, 179, 180, 181, 150, 151, 152, 154, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 62, 63, 64, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 127, 128, 129, 102, 103, 104, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 80, 81, 82, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 133, 134, 135, 137, 139, 140, 141, 142, 143, 144, 145, 146], "/Users/flavianshem/Projects/shopping_list/app/tests/test_user.py": [1, 2, 3, 4, 5, 6, 7, 10, 11, 12, 21, 25, 31, 39, 51, 61, 71, 86, 13, 14, 15, 16, 17, 18, 19, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 22, 23, 88, 89, 90, 91, 92, 93, 94, 95, 27, 28, 29, 33, 34, 35, 36, 37, 41, 42, 43, 44, 45, 46, 47, 48, 49, 53, 54, 55, 56, 57, 58, 59, 63, 64, 65, 66, 67, 68, 69], "/Users/flavianshem/Projects/shopping_list/manage.py": [], "/Users/flavianshem/Projects/shopping_list/run.py": []}}
