+++ .gitignore
@@ -2,4 +2,5 @@
 .vscode
 *.pyc
 __pycache__/
-
+htmlcov/
+.DS_Store
