+++ .travis.yml
@@ -12,6 +12,6 @@ before_script:
   - psql -c 'create database shopping_list_tdb1;' -U postgres
 # command to run tests
 script:
-  - python -m pytest --cov tests
+  - python -m pytest --cov=. app/tests
 after_success:
   - coveralls
